requirements.txt
pandas
numpy
selenium
matplotlib
redis
mysql-connector-python
openpyxl